fix(chaining): fix review

+++ openaev-api/src/main/java/io/openaev/service/chaining/WorkflowService.java
@@ -998,6 +998,7 @@ public class WorkflowService {
   }
 
   private void deleteAllAssetAgentJobs(List<String> injectsIds, String tenantId) {
+    if (CollectionUtils.isEmpty(injectsIds)) return;
     assetAgentJobRepository.deleteAllByInjectIdsAndTenantId(injectsIds, tenantId);
   }
